refactor(design-system): split bacteria scenes by domain
